Created new utilities library. Moved the reading average function into new file.
